+++ pkg/httpx/httpx.go
@@ -1,0 +1,7 @@
+package httpx
+
+const (
+	HeaderContentType   = "Content-Type"
+	HeaderContentLength = "Content-Length"
+	HeaderAcceptRanges  = "Accept-Ranges"
+)