+++ pkg/httpx/metrics.go
@@ -1,4 +1,4 @@
-package mux
+package httpx
 
 import "github.com/prometheus/client_golang/prometheus"