+++ pkg/oci/client.go
@@ -4,12 +4,10 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"io"
 	"net/http"
 	"net/url"
 	"runtime"
-	"slices"
 	"strconv"
 	"strings"
 	"sync"
@@ -18,37 +16,13 @@ import (
 	"github.com/containerd/containerd/v2/core/images"
 	"github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+
+	"github.com/spegel-org/spegel/pkg/httpx"
 )
 
 const (
-	DigestHeader        = "Docker-Content-Digest"
-	ContentTypeHeader   = "Content-Type"
-	ContentLengthHeader = "Content-Length"
+	HeaderDockerDigest = "Docker-Content-Digest"
 )
-
-type StatusError struct {
-	Content    string
-	StatusCode int
-}
-
-func (e *StatusError) Error() string {
-	return fmt.Sprintf("unexpected status code %d with body %s", e.StatusCode, e.Content)
-}
-
-func CheckResponseStatus(resp *http.Response, expected ...int) error {
-	if slices.Contains(expected, resp.StatusCode) {
-		return nil
-	}
-	defer resp.Body.Close()
-	b, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-	return &StatusError{
-		StatusCode: resp.StatusCode,
-		Content:    string(b),
-	}
-}
 
 type Client struct {
 	hc *http.Client
@@ -226,38 +200,14 @@ func (c *Client) fetch(ctx context.Context, method string, dist DistributionPath
 			c.tc.Store(tcKey, token)
 			continue
 		}
-		err = CheckResponseStatus(resp, http.StatusOK)
+		err = httpx.CheckResponseStatus(resp, http.StatusOK, http.StatusPartialContent)
 		if err != nil {
 			return nil, ocispec.Descriptor{}, err
 		}
 
-		dgst := dist.Digest
-		dgstStr := resp.Header.Get(DigestHeader)
-		if dgstStr != "" {
-			dgst, err = digest.Parse(dgstStr)
-			if err != nil {
-				return nil, ocispec.Descriptor{}, err
-			}
-		}
-		if dgst == "" {
-			return nil, ocispec.Descriptor{}, errors.New("digest cannot be empty")
-		}
-		mt := resp.Header.Get(ContentTypeHeader)
-		if mt == "" {
-			return nil, ocispec.Descriptor{}, errors.New("content type header cannot be empty")
-		}
-		cl := resp.Header.Get(ContentLengthHeader)
-		if cl == "" {
-			return nil, ocispec.Descriptor{}, errors.New("content length header cannot be empty")
-		}
-		size, err := strconv.ParseInt(cl, 10, 64)
+		desc, err := DescriptorFromHeader(resp.Header)
 		if err != nil {
 			return nil, ocispec.Descriptor{}, err
-		}
-		desc := ocispec.Descriptor{
-			Digest:    dgst,
-			MediaType: mt,
-			Size:      size,
 		}
 		return resp.Body, desc, nil
 	}
@@ -297,7 +247,7 @@ func getBearerToken(ctx context.Context, wwwAuth string, client *http.Client) (s
 	if err != nil {
 		return "", err
 	}
-	err = CheckResponseStatus(resp, http.StatusOK)
+	err = httpx.CheckResponseStatus(resp, http.StatusOK)
 	if err != nil {
 		return "", err
 	}
@@ -314,4 +264,35 @@ func getBearerToken(ctx context.Context, wwwAuth string, client *http.Client) (s
 		return "", err
 	}
 	return tokenResp.Token, nil
+}
+
+func DescriptorFromHeader(header http.Header) (ocispec.Descriptor, error) {
+	mediaType := header.Get(httpx.HeaderContentType)
+	if mediaType == "" {
+		return ocispec.Descriptor{}, errors.New("content type cannot be empty")
+	}
+	contentLength := header.Get(httpx.HeaderContentLength)
+	if contentLength == "" {
+		return ocispec.Descriptor{}, errors.New("content length cannot be empty")
+	}
+	size, err := strconv.ParseInt(contentLength, 10, 64)
+	if err != nil {
+		return ocispec.Descriptor{}, err
+	}
+	dgst, err := digest.Parse(header.Get(HeaderDockerDigest))
+	if err != nil {
+		return ocispec.Descriptor{}, err
+	}
+	desc := ocispec.Descriptor{
+		MediaType: mediaType,
+		Size:      size,
+		Digest:    dgst,
+	}
+	return desc, nil
+}
+
+func WriteDescriptorToHeader(desc ocispec.Descriptor, header http.Header) {
+	header.Set(httpx.HeaderContentType, desc.MediaType)
+	header.Set(httpx.HeaderContentLength, strconv.FormatInt(desc.Size, 10))
+	header.Set(HeaderDockerDigest, desc.Digest.String())
 }